+++ src/page/article.js
@@ -8,7 +8,6 @@
  */
 import comArticle from "./common/com-article";
 import imgBox from "../components/imgBox/imgBox";
-await $.__tools.dynamicLoadingJs($.__config.default.iconfont).catch(e => console.error('iconfont.js', e))
 
 export default function main() {