+++ docs/v2/Docs/changelog.md
@@ -2,6 +2,7 @@
 - 修复diggit的配置项
 - 修复非文章页更新图表显示异常问题
 - 优化评论打字特效，调整配置`articleContent.commentTyping`
+- 文档更新
 
 # 2022.2.15 - v2.2.7
 - 新增评论打字特效配置`articleContent.commentTyping`